pyproject.toml
[tool.poetry]
name = "tokyo-travel-guide"
version = "0.1.0"
description = "Interactive Features for my Tokyo Travel Guide"
authors = ["Joe Gambino <joe@gam.bingo>"]
readme = "README.md"
packages = [{include = "tokyo_travel_guide"}]

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.25.0"
osmnx = "^1.5.1"
scikit-learn = "^1.3.0"
pandas = "^2.0.3"
altair = "^5.0.1"
pyyaml = "^6.0.1"
streamlit-folium = "^0.13.0"
tqdm = "^4.65.0"


[tool.poetry.group.dev.dependencies]
watchdog = "^3.0.0"
matplotlib = "^3.7.2"
jupyterlab = "^4.0.3"
ipykernel = "^6.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
